+++ commands/src/main/java/org/wildfly/extras/creaper/commands/datasources/AddXADataSource.java
@@ -10,9 +10,9 @@ import org.wildfly.extras.creaper.core.online.OnlineCommand;
 import org.wildfly.extras.creaper.core.online.OnlineCommandContext;
 import org.wildfly.extras.creaper.core.online.operations.Address;
 import org.wildfly.extras.creaper.core.online.operations.Batch;
-import org.wildfly.extras.creaper.core.online.operations.OperationException;
 import org.wildfly.extras.creaper.core.online.operations.Operations;
 import org.wildfly.extras.creaper.core.online.operations.Values;
+import org.wildfly.extras.creaper.core.online.operations.admin.Administration;
 
 import java.io.IOException;
 import java.util.HashMap;
@@ -173,7 +173,8 @@ public class AddXADataSource implements OnlineCommand, OfflineCommand {
         if (replaceExisting) {
             try {
                 ops.removeIfExists(dsAddress);
-            } catch (OperationException e) {
+                new Administration(ctx.client).reloadIfRequired();
+            } catch (Exception e) {
                 throw new CommandFailedException("Failed to remove existing XA datasource " + name, e);
             }
         }
@@ -453,7 +454,7 @@ public class AddXADataSource implements OnlineCommand, OfflineCommand {
 
         /**
          * If xa datasource with the given pool-name exists it will be replaced
-         * with this newly created datasource.
+         * with this newly created datasource. <b>Note that when enabled, this can cause server reload!</b>
          */
         public final THIS replaceExisting() {
             this.replaceExisting = true;